Added a build system with cmake
Prior to this commit it wasn't actually possible to build the project.
This commit introduces CMakeLists.txt to aid compilation. Compilation
instructions were also added to the README.

With this change, configuration was moved from crisp8/config.h to cmake,
so that file was no longer needed and was therefore removed.