Added stack implementation
This commit implements the previously commited stack api in
crisp8/stack.c.

To facilitate future file additions, CMakeLists.txt was
updated to simply use all source files in the crisp8 directory.

Furthermore, an examples directory was added to test features, as
well as to inform developers of how to use the library.