Move the chip-8 struct to header file
Previously the chip8_s struct lived in crisp8.c to provide
encapsulation to the user of the library, however this posed problems
when splitting functions into multiple files. As a solution, the struct
was moved to its own private header file, only to be included by non
public API headers.